+++ packages/components/src/components/Dialog/styles/dialogStyle.ts
